src/third_party/WebKit/LayoutTests/external/wpt/fetch/sec-metadata/embed.tentative.https.sub-expected.txt

This is a testharness.js-based test.
FAIL Same-Origin embed assert_equals: cause expected (undefined) undefined but got (string) "forced"
FAIL Same-Site embed assert_equals: cause expected (undefined) undefined but got (string) "forced"
FAIL Cross-Site embed assert_equals: cause expected (undefined) undefined but got (string) "forced"
Harness: the test ran to completion.
